# openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/products": {
			"post": {
				"operationId": "Products_create",
				"parameters": [ ],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductDto"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"get": {
				"operationId": "Products_findAll",
				"parameters": [
					{
						"required": false,
						"name": "attributeValueIds",
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"type": "number"
							}
						}
					},
					{
						"required": false,
						"name": "minPrice",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"required": false,
						"name": "maxPrice",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"required": false,
						"name": "sorting",
						"in": "query",
						"schema": {
							"$ref": "#/components/schemas/ProductSorting"
						}
					},
					{
						"required": false,
						"name": "language",
						"in": "query",
						"schema": {
							"$ref": "#/components/schemas/LanguageCode"
						}
					},
					{
						"required": false,
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"required": false,
						"name": "variants",
						"in": "query",
						"schema": {
							"type": "boolean"
						}
					},
					{
						"required": false,
						"name": "itemsPerPage",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"required": false,
						"name": "page",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"required": false,
						"name": "categoryId",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"required": false,
						"name": "collectionId",
						"in": "query",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductFilterResultDto"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/products/ids": {
			"get": {
				"operationId": "Products_getAllIds",
				"parameters": [ ],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "number"
									}
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/products/{id}": {
			"get": {
				"operationId": "Products_getById",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductDto"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "Products_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductDto"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "Products_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductDto"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/products/{productId}/variants": {
			"post": {
				"operationId": "ProductVariant_create",
				"parameters": [
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductVariantDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductVariantDto"
								}
							}
						}
					}
				},
				"tags": [
					"Product Variants"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"get": {
				"operationId": "ProductVariant_findOfProduct",
				"parameters": [
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ProductVariantDto"
									}
								}
							}
						}
					}
				},
				"tags": [
					"Product Variants"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/products/{productId}/variants/{id}": {
			"get": {
				"operationId": "ProductVariant_getById",
				"parameters": [
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductVariantDto"
								}
							}
						}
					}
				},
				"tags": [
					"Product Variants"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "ProductVariant_update",
				"parameters": [
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductVariantDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductVariantDto"
								}
							}
						}
					}
				},
				"tags": [
					"Product Variants"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "ProductVariant_remove",
				"parameters": [
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductVariantDto"
								}
							}
						}
					}
				},
				"tags": [
					"Product Variants"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
		"/api/v1/discounts/{id}/products/{productId}": {
			"delete": {
				"operationId": "DiscountSales_removeProduct",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/variants/{variantId}": {
			"delete": {
				"operationId": "DiscountSales_removeVariant",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "variantId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
		"/api/v1/collections/{collectionId}/products/{productId}": {
			"delete": {
				"operationId": "Collections_removeProduct",
				"parameters": [
					{
						"name": "collectionId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Collections"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
	},
	"info": {
		"title": "Nima CMS",
		"description": "The core api for the Nima CMS backend.",
		"version": "0.0.1-alpha",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "http://localhost:3333",
			"description": "Localhost"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"TranslatableDto": {
				"type": "object",
				"properties": {
					"de": {
						"type": "string"
					},
					"it": {
						"type": "string"
					},
					"fr": {
						"type": "string"
					},
					"es": {
						"type": "string"
					},
					"el": {
						"type": "string"
					},
					"en": {
						"type": "string"
					}
				}
			},
...
			"CreateAssignedProductAttributeDto": {
				"type": "object",
				"properties": {
					"productTypeAttributeId": {
						"type": "number"
					},
					"values": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateAssignedProductAttributeValueDto"
						}
					}
				},
				"required": [
					"productTypeAttributeId",
					"values"
				]
			},
			"CreateSortableMediaDto": {
				"type": "object",
				"properties": {
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"mediaId": {
						"type": "number"
					}
				},
				"required": [
					"sortOrder",
					"mediaId"
				]
			},
			"CreateProductDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Product Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Product Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"type": "string",
						"example": "SEO Description"
					},
					"seoTitle": {
						"type": "string",
						"example": "SEO Title"
					},
					"chargeTaxes": {
						"type": "boolean",
						"example": true
					},
					"weight": {
						"type": "number",
						"example": 12.3
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"slug": {
						"type": "string",
						"example": "product-name"
					},
					"descriptionPlaintext": {
						"type": "string",
						"example": "Product Description"
					},
					"rating": {
						"type": "number",
						"example": 10
					},
					"isPublished": {
						"type": "boolean",
						"example": true
					},
					"isVisibleInListings": {
						"type": "boolean",
						"example": true
					},
					"isAvailableForPurchase": {
						"type": "boolean",
						"example": true
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"minPrice": {
						"type": "number",
						"example": 12.3
					},
					"productTypeId": {
						"type": "number"
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateAssignedProductAttributeDto"
						}
					},
					"categoryId": {
						"type": "number"
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateSortableMediaDto"
						}
					},
					"collectionIds": {
						"type": "array",
						"items": {
							"type": "number"
						}
					}
				},
				"required": [
					"name",
					"description",
					"seoDescription",
					"seoTitle",
					"chargeTaxes",
					"weight",
					"metadata",
					"privateMetadata",
					"slug",
					"descriptionPlaintext",
					"rating",
					"isPublished",
					"isVisibleInListings",
					"isAvailableForPurchase",
					"currency",
					"minPrice",
					"productTypeId",
					"attributes",
					"categoryId",
					"productMedia",
					"collectionIds"
				]
			},
			"ProductAttributeDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number"
					},
					"name": {
						"example": {
							"en": "Attribute Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"availableInGrid": {
						"type": "boolean",
						"example": true
					},
					"visibleInStorefront": {
						"type": "boolean",
						"example": true
					},
					"filterableInDashboard": {
						"type": "boolean",
						"example": true
					},
					"filterableInStorefront": {
						"type": "boolean",
						"example": true
					},
					"valueRequired": {
						"type": "boolean",
						"example": true
					},
					"storefrontSearchPosition": {
						"type": "number",
						"example": 0
					},
					"inputType": {
						"example": "DROPDOWN",
						"$ref": "#/components/schemas/InputType"
					},
					"unit": {
						"$ref": "#/components/schemas/Unit"
					},
					"values": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AttributeValueDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"metadata",
					"privateMetadata",
					"availableInGrid",
					"visibleInStorefront",
					"filterableInDashboard",
					"filterableInStorefront",
					"valueRequired",
					"storefrontSearchPosition",
					"inputType",
					"values"
				]
			},
...
			"SortableMediaDto": {
				"type": "object",
				"properties": {
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"media": {
						"$ref": "#/components/schemas/MediaDto"
					}
				},
				"required": [
					"sortOrder",
					"media"
				]
			},
			"ProductVariantDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"sku": {
						"type": "string",
						"example": "product-sku-1"
					},
					"name": {
						"example": {
							"en": "Product Variant Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"weight": {
						"type": "number",
						"example": 1
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"created": {
						"type": "string",
						"example": "2022-01-01T00:00:00.000+0200"
					},
					"updatedAt": {
						"type": "string",
						"example": "2022-01-01T00:00:00.000+0200"
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"priceAmount": {
						"type": "number",
						"example": 1.4
					},
					"discountedPrice": {
						"type": "number",
						"example": 1.4
					},
					"stock": {
						"type": "number",
						"example": 10
					},
					"trackInventory": {
						"type": "boolean",
						"example": true
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductAttributeDto"
						}
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SortableMediaDto"
						}
					}
				},
				"required": [
					"id",
					"sku",
					"name",
					"metadata",
					"privateMetadata",
					"created",
					"updatedAt",
					"currency",
					"stock",
					"trackInventory",
					"attributes",
					"productMedia"
				]
			},
			"ProductDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Product Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Product Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"updatedAt": {
						"type": "string",
						"example": ""
					},
					"seoDescription": {
						"type": "string",
						"example": "SEO Description"
					},
					"seoTitle": {
						"type": "string",
						"example": "SEO Title"
					},
					"chargeTaxes": {
						"type": "boolean",
						"example": true
					},
					"weight": {
						"type": "number",
						"example": 12.3
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"slug": {
						"type": "string",
						"example": "product-name"
					},
					"defaultVariantId": {
						"type": "number"
					},
					"descriptionPlaintext": {
						"type": "string",
						"example": "Product Description"
					},
					"rating": {
						"type": "number",
						"example": 10
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"isPublished": {
						"type": "boolean",
						"example": true
					},
					"isVisibleInListings": {
						"type": "boolean",
						"example": true
					},
					"isAvailableForPurchase": {
						"type": "boolean",
						"example": true
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"minPrice": {
						"type": "number",
						"example": 12.3
					},
					"productTypeId": {
						"type": "number"
					},
					"categoryId": {
						"type": "number"
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductAttributeDto"
						}
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SortableMediaDto"
						}
					},
					"collections": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionDto"
						}
					},
					"defaultVariant": {
						"$ref": "#/components/schemas/ProductVariantDto"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"updatedAt",
					"seoDescription",
					"seoTitle",
					"chargeTaxes",
					"weight",
					"metadata",
					"privateMetadata",
					"slug",
					"defaultVariantId",
					"descriptionPlaintext",
					"rating",
					"created",
					"isPublished",
					"isVisibleInListings",
					"isAvailableForPurchase",
					"currency",
					"minPrice",
					"productTypeId",
					"categoryId",
					"attributes",
					"productMedia",
					"collections"
				]
			},
			"CollectionProductsDto": {
				"type": "object",
				"properties": {
					"sortOrder": {
						"type": "number"
					},
					"product": {
						"$ref": "#/components/schemas/ProductDto"
					}
				},
				"required": [
					"sortOrder",
					"product"
				]
			},
			"CollectionDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Collection Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "collection-name"
					},
					"seoDescription": {
						"example": "SEO Description",
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoTitle": {
						"example": "SEO Title",
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Collection Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"backgroundImage": {
						"$ref": "#/components/schemas/MediaDto"
					},
					"products": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionProductsDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"seoDescription",
					"seoTitle",
					"description",
					"metadata",
					"privateMetadata",
					"backgroundImage",
					"products"
				]
			},
			"ProductSorting": {
				"type": "string",
				"enum": [
					"NAME_ASC",
					"NAME_DESC",
					"PRICE_ASC",
					"PRICE_DESC",
					"DATE_CREATED_ASC",
					"DATE_CREATED_DESC",
					"RATING_ASC",
					"RATING_DESC"
				]
			},
			"LanguageCode": {
				"type": "string",
				"enum": [
					"en",
					"el",
					"fr",
					"es",
					"it",
					"de",
					"ru"
				]
			},
...
			"AttributeDrillDownDto": {
				"type": "object",
				"properties": {
					"attributeId": {
						"type": "number"
					},
					"attributeSlug": {
						"type": "string"
					},
					"fieldValues": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AttributeDrillDownValuesDto"
						}
					}
				},
				"required": [
					"attributeId",
					"attributeSlug",
					"fieldValues"
				]
			},
			"ProductFilterResultDto": {
				"type": "object",
				"properties": {
					"attributeDrillDown": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AttributeDrillDownDto"
						}
					},
					"minPrice": {
						"type": "number"
					},
					"maxPrice": {
						"type": "number"
					},
					"selectedMinPrice": {
						"type": "number"
					},
					"selectedMaxPrice": {
						"type": "number"
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductDto"
						}
					},
					"pageNumber": {
						"type": "number"
					},
					"pageSize": {
						"type": "number"
					},
					"totalCount": {
						"type": "number"
					}
				},
				"required": [
					"attributeDrillDown",
					"minPrice",
					"maxPrice",
					"selectedMinPrice",
					"selectedMaxPrice",
					"items",
					"pageNumber",
					"pageSize",
					"totalCount"
				]
			},
...
			"CreateAssignedProductVariantAttributeDto": {
				"type": "object",
				"properties": {
					"productTypeVariantAttributeId": {
						"type": "number"
					},
					"values": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateAssignedProductVariantAttributeValueDto"
						}
					}
				},
				"required": [
					"productTypeVariantAttributeId",
					"values"
				]
			},
			"CreateProductVariantDto": {
				"type": "object",
				"properties": {
					"sku": {
						"type": "string",
						"example": "product-sku-1"
					},
					"name": {
						"example": {
							"en": "Product Variant Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"weight": {
						"type": "number",
						"example": 1
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"priceAmount": {
						"type": "number",
						"example": 1.4
					},
					"stock": {
						"type": "number",
						"example": 10
					},
					"trackInventory": {
						"type": "boolean",
						"example": true
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateAssignedProductVariantAttributeDto"
						}
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateSortableMediaDto"
						}
					}
				},
				"required": [
					"sku",
					"name",
					"metadata",
					"privateMetadata",
					"currency",
					"stock",
					"trackInventory",
					"attributes",
					"productMedia"
				]
			},
...
			"CategoryDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"seoTitle": {
						"example": {
							"en": "Category Seo Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"example": {
							"en": "Category Seo Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"name": {
						"example": {
							"en": "Category Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Category Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "category-name"
					},
					"children": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"parent": {
						"description": "Only for ancestors query",
						"allOf": [
							{
								"$ref": "#/components/schemas/CategoryDto"
							}
						]
					}
				},
				"required": [
					"id",
					"privateMetadata",
					"metadata",
					"seoTitle",
					"seoDescription",
					"name",
					"description",
					"slug",
					"children"
				]
			},
...
			"DiscountType": {
				"type": "string",
				"enum": [
					"PERCENTAGE",
					"FLAT"
				]
			},
...
			"DiscountSaleDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Discount Sale Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"discountType": {
						"$ref": "#/components/schemas/DiscountType"
					},
					"discountValue": {
						"type": "number"
					},
					"endDate": {
						"type": "string",
						"example": "2022-12-31"
					},
					"startDate": {
						"type": "string",
						"example": "2022-01-01"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"updatedAt": {
						"type": "string",
						"example": "2022-01-01"
					},
					"categories": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"products": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductDto"
						}
					},
					"variants": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductVariantDto"
						}
					},
					"collections": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"discountType",
					"discountValue",
					"metadata",
					"privateMetadata",
					"created",
					"updatedAt",
					"categories",
					"products",
					"variants",
					"collections"
				]
			},
...
		}
	}
}